ReaderMiddleware added
