+++ dtslint/ts3.5/index.ts
@@ -1,13 +1,13 @@
-import * as H from '../../src'
+import { connection as H, middleware as HM } from '../../src'
 import { pipe } from 'fp-ts/lib/pipeable'
 
 // $ExpectError
-const m1 = H.status(1000)
+const m1 = HM.status(1000)
 
 const m2 = pipe(
-  H.status(H.Status.OK),
-  H.ichain(() => H.closeHeaders()),
-  H.ichain(() => H.send('Hello hyper-ts!')),
+  HM.status(H.Status.OK),
+  HM.ichain(() => HM.closeHeaders()),
+  HM.ichain(() => HM.send('Hello hyper-ts!')),
   // $ExpectError
-  H.ichain(() => () => H.header('field', 'value'))
+  HM.ichain(() => () => HM.header('field', 'value'))
 )
