+++ examples/decodeBody.ts
@@ -1,5 +1,5 @@
-import * as H from '../src'
+import { middleware as HM } from '../src'
 import * as t from 'io-ts'
 
 // return a middleware validating `req.body`
-export const middleware = H.decodeBody(t.string.decode)
+export const middleware = HM.decodeBody(t.string.decode)
