+++ examples/decodeParams.ts
@@ -1,8 +1,8 @@
-import * as H from '../src'
+import { middleware as HM } from '../src'
 import * as t from 'io-ts'
 
 // returns a middleware validating both `req.param.user_id` and `req.param.user_name`
-export const middleware = H.decodeParams(
+export const middleware = HM.decodeParams(
   t.strict({
     user_id: t.string,
     user_name: t.string
