+++ examples/decodeQuery.ts
@@ -1,8 +1,8 @@
-import * as H from '../src'
+import { middleware as HM } from '../src'
 import * as t from 'io-ts'
 
 // return a middleware validating the query "order=desc&shoe[color]=blue&shoe[type]=converse"
-export const middleware = H.decodeQuery(
+export const middleware = HM.decodeQuery(
   t.strict({
     order: t.string,
     shoe: t.strict({
